Extracted state and event classes to separate files
