+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/connect/CardReaderConnectEvent.kt
@@ -1,0 +1,34 @@
+package com.woocommerce.android.ui.prefs.cardreader.connect
+
+import androidx.annotation.StringRes
+import com.woocommerce.android.cardreader.CardReaderManager
+import com.woocommerce.android.viewmodel.MultiLiveEvent
+
+sealed class CardReaderConnectEvent : MultiLiveEvent.Event() {
+    data class InitializeCardReaderManager(val onCardManagerInitialized: (manager: CardReaderManager) -> Unit) :
+        CardReaderConnectEvent()
+
+    data class CheckLocationPermissions(val onPermissionsCheckResult: (Boolean) -> Unit) : CardReaderConnectEvent()
+
+    data class CheckLocationEnabled(val onLocationEnabledCheckResult: (Boolean) -> Unit) : CardReaderConnectEvent()
+
+    data class CheckBluetoothEnabled(val onBluetoothCheckResult: (Boolean) -> Unit) : CardReaderConnectEvent()
+
+    data class RequestEnableBluetooth(val onEnableBluetoothRequestResult: (Boolean) -> Unit) :
+        CardReaderConnectEvent()
+
+    data class RequestLocationPermissions(val onPermissionsRequestResult: (Boolean) -> Unit) :
+        CardReaderConnectEvent()
+
+    object OpenPermissionsSettings : CardReaderConnectEvent()
+
+    data class OpenLocationSettings(val onLocationSettingsClosed: () -> Unit) : CardReaderConnectEvent()
+
+    object ShowCardReaderTutorial : CardReaderConnectEvent()
+
+    object ShowUpdateInProgress : CardReaderConnectEvent()
+
+    object NavigateToOnboardingFlow : CardReaderConnectEvent()
+
+    data class ShowToast(@StringRes val message: Int) : CardReaderConnectEvent()
+}
